Fix server side caching bug (#649)
* Fix caching bug in server
* Show user ID in title

+++ api/Engraved.Core/Source/Application/Commands/Metrics/Delete/DeleteMetricCommandExecutor.cs
@@ -1,4 +1,5 @@
 ﻿using Engraved.Core.Application.Persistence;
+using Engraved.Core.Domain.Metrics;
 
 namespace Engraved.Core.Application.Commands.Metrics.Delete;
 
@@ -13,8 +14,14 @@ public class DeleteMetricCommandExecutor : ICommandExecutor
 
   public async Task<CommandResult> Execute(IRepository repository, IDateService dateService)
   {
+    IMetric? metric = await repository.GetMetric(_command.Id);
+    if (metric == null)
+    {
+      return new CommandResult();
+    }
+
     await repository.DeleteMetric(_command.Id);
 
-    return new CommandResult { EntityId = _command.Id };
+    return new CommandResult(_command.Id, metric.Permissions.GetUserIdsWithAccess());
   }
 }
